added favicon, updated landing page routes
