+++ Webapp/sources/blueprints/reaction_list/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+reaction_list_bp = Blueprint("reaction_list", __name__)
+
+from . import routes
