--- Webapp/sources/blueprints/standard_landing_page/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-standard_landing_page_bp = Blueprint("standard_landing_page", __name__)
-
-from . import routes
