+++ Webapp/sources/templates/landing_page.html
@@ -1,0 +1,11 @@
+<head>
+   <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='css/landing_page.css') }}">
+</head>
+{% extends "base.html" %}
+{% block content %}
+
+<p id="bg" class="m-3 p-3">
+
+</p>
+
+{% endblock %}
