+++ Webapp/sources/templates/manage_account.html
@@ -7,7 +7,7 @@
         <p>Username: {{ current_user.username }}</p>
         <p>Email: {{ current_user.email }}</p>
         <p><a href="{{ url_for('update_email.update_email_password') }}" id="update-email-password" class="btn btn-success">Change Email or Password</a></p>
-        <p><a href="{{ url_for('standard_landing_page.delete_profile') }}" id="delete-profile" class="btn btn-primary">Delete Profile</a></p>
+        <p><a href="{{ url_for('delete_profile.delete_profile') }}" id="delete-profile" class="btn btn-primary">Delete Profile</a></p>
     </div>
 </div>
 </body>
